rename application to glortho
