+++ alembic.ini
@@ -2,11 +2,11 @@
 
 [alembic]
 # path to migration scripts
-script_location = ./jwt_authentication_python_postgres/db/migrations
+script_location = ./glortho/db/migrations
 # version location specification; this defaults
 # to alembic/versions.  When using multiple version
 # directories, initial revisions must be specified with --version-path
-version_locations = ./jwt_authentication_python_postgres/db/migrations/versions
+version_locations = ./glortho/db/migrations/versions
 
 [loggers]
 keys = root,sqlalchemy,alembic
