+++ docker-compose.test.yml
@@ -9,7 +9,7 @@ services:
       POSTGRES_DB: test_db
 
   sut:
-    image: 'jwt_authentication_python_postgres:test'
+    image: 'glortho:test'
     build:
       context: .
       target: development
