+++ docker-compose.yml
@@ -9,7 +9,7 @@ services:
       - .dockerenv
 
   app:
-    image: 'jwt_authentication_python_postgres:dev'
+    image: 'glortho:dev'
     build: .
     ports:
       - 8082:8080
