+++ glortho/api/dependencies/user.py
@@ -2,12 +2,12 @@ from fastapi import Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordBearer
 from jose import JWTError, jwt
 
-from jwt_authentication_python_postgres.api.dependencies.database import get_repository
-from jwt_authentication_python_postgres.api.dependencies.settings import get_settings
-from jwt_authentication_python_postgres.core.config import Settings
-from jwt_authentication_python_postgres.db.repositories.users import UsersRepository
-from jwt_authentication_python_postgres.models.token import TokenData
-from jwt_authentication_python_postgres.models.user import User
+from glortho.api.dependencies.database import get_repository
+from glortho.api.dependencies.settings import get_settings
+from glortho.core.config import Settings
+from glortho.db.repositories.users import UsersRepository
+from glortho.models.token import TokenData
+from glortho.models.user import User
 
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
