+++ glortho/api/routes/__init__.py
@@ -1,8 +1,8 @@
 from fastapi import APIRouter
-from jwt_authentication_python_postgres.api.routes.cleanings import (
+from glortho.api.routes.cleanings import (
     router as cleanings_router,
 )
-from jwt_authentication_python_postgres.api.routes.users import (
+from glortho.api.routes.users import (
     router as users_router,
 )
