+++ glortho/models/user.py
@@ -1,6 +1,6 @@
 from email_validator import validate_email, EmailNotValidError
 from typing import Optional
-from jwt_authentication_python_postgres.models.core import IDModelMixin, CoreModel
+from glortho.models.core import IDModelMixin, CoreModel
 from pydantic import constr, validator
