+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
-name = "jwt-authentication-python-postgres"
-version = "0.1.0"
+name = "glortho"
+version = "0.2.0"
 description = ""
 authors = ["Andrew Hoetker <andrew@hoetker.engineer>"]
