--- tests/test_jwt_authentication_python_postgres.py
@@ -1,5 +1,0 @@
-from jwt_authentication_python_postgres import __version__
-
-
-def test_version():
-    assert __version__ == "0.1.0"
